+++ users/urls.py
@@ -6,5 +6,4 @@ urlpatterns = [
     path("/signup", SignupView.as_view()),
     path("/login", LoginView.as_view()),
     path("/detail", UserDetailView.as_view()),
-    path("/like", LikeView.as_view())
 ]
